Fix a common bug with do-notation
Because of the way iterators are "depleted" upon use,
the returned Future could only safely be forked once.

This is something which is known within the community
to be unfixable in a general way. However, because
Future.do is specialized, I was able to fix it by
respawning the iterator upon every new fork().