Remove module declaration from TypeScript typings
I believe this was added by mistake when the TypeScript types were
introduced. I think it was probably copy pasted over from Tushar's
personal TypeScript overrides for the 'fluture' module.

It never caused any problems, because users have always been able to
import Fluture from the 'fluture' module. However, this changes with
Deno's dependency model.

This commit removes the module declaration wrapper around the TypeScript
types, forcing TypeScript to make the association with the relevant
module in a different way. In particular, TypeScript should now make the
association based either on the location of the file (when imported
using Node's require algorithm), or based on a <reference> directive,
such as the one used in dist/module.js.