Relax the TypeScript type for alt and add branch awareness
The type for 'alt' did not allow users to supply a future with a
predetermined branch (when one of the branches is never).

This commit adds overloads for alt that allow TypeScript to understand
how our code recovers when there is no concrete type in either of the
branches of the first Future. This makes the type inference much better,
but it also means that mixed types are no longer considered a type error.

Finally, this commit adds tests for the TypeScript type of alt.